+++ client/proxmark3.h
@@ -13,6 +13,7 @@
 #define PROXMARK3_H__
 
 #include "usb_cmd.h"
+#include "cmdscript.h"  // CmdScriptRun  
 
 #define PROXPROMPT "pm3 --> "
 #ifdef __cplusplus
@@ -22,7 +23,7 @@ extern "C" {
 void SendCommand(UsbCommand *c);
 const char *get_my_executable_path(void);
 const char *get_my_executable_directory(void);
-void main_loop(char *script_cmds_file, bool usb_present);
+void main_loop(char *script_cmds_file, char *script_cmd, bool usb_present);
 
 #ifdef __cplusplus
 }