+++ client/ui.h
@@ -42,6 +42,8 @@ extern int flushAfterWrite;   //buzzy
 extern bool GridLocked;
 extern bool showDemod;
 
+extern pthread_mutex_t print_lock;
+
 extern void iceIIR_Butterworth(int * data, const size_t len);
 extern void iceSimple_Filter(int *data, const size_t len, uint8_t k);
 #endif